+++ src/elements/db/ElementQuery.php
@@ -1127,7 +1127,11 @@ class ElementQuery extends Query implements ElementQueryInterface
      */
     public function andRelatedTo($value)
     {
-        if ($this->relatedTo === null) {
+        if (!$value) {
+            return $this;
+        }
+
+        if (!$this->relatedTo) {
             return $this->relatedTo($value);
         }